Docker file and scripts changes for Dotnet (Linux+Windows) and Java(Currently only linux, Windows will be uploaded soon).

+++ enterprise/dotnet edition/client/Linux/resources/installncache.sh
@@ -37,6 +37,10 @@ while [ "$1" != "" ]; do
 			EMAIL=$1
 			;;
 			
+	-g | --installplayground )	shift
+			PLAYGROUND=$1
+			;;
+			
 	-v | --verbose )
 			VERBOSE="true"
 			;;
@@ -62,22 +66,29 @@ then
 	PASSWORD="ncache"
 fi
 
+setup_file="ncache.ent.net.tar.gz"
+
+if [ ! -e "$setup_file"  ]; then
+    echo "Error: $setup_file does not exist."
+	exit 1
+fi
+
 #--- Untaring and installing NCache
-tar -zxf ncache.ent.net.tar.gz
+tar -zxf $setup_file
+cd ncache-enterprise
 cd ncache-enterprise
 
 ./install --firstname $FIRST_NAME --lastname $LAST_NAME --email $EMAIL --company $COMPANY --installmode $INSTALLMODE
 
 cd ..
 #--- Removing installation resources
-rm -f ncache.ent.net.tar.gz
+rm -f $setup_file
 rm -rf ncache-enterprise/
 
-mkdir /home/ncache
-
 #--- Updating permissions and ownership
-chmod -R 775 /home/ncache /opt/ncache/bin/service /opt/ncache/libexec 
-chown -R ncache:root /app /opt/ncache /home/ncache
+chmod -R 775 /home/ncache /opt/ncache/bin/tools/web /opt/ncache/bin/service /opt/ncache/libexec 
+chown -R ncache:root /app
+chown -R ncache:ncache /opt/ncache
 
 #--- Adding the user to the root group
 usermod -a -G root ncache
